+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Next release
+* Add low speed support for roller shutters
+* roller shutter position is no more a property
 ### 0.2.2
 * Add orientation support for blind
 ### 0.2.1